+++ packages/react/src/components/format/format.stories.tsx
@@ -11,9 +11,13 @@ export { ByteSizes } from './examples/byte-sizes'
 export { ByteWithLocale } from './examples/byte-with-locale'
 export { ByteWithUnit } from './examples/byte-with-unit'
 export { ByteWithUnitDisplay } from './examples/byte-with-unit-display'
+
 export { NumberBasic } from './examples/number-basic'
 export { NumberWithCompact } from './examples/number-with-compact'
 export { NumberWithCurrency } from './examples/number-with-currency'
 export { NumberWithLocale } from './examples/number-with-locale'
 export { NumberWithPercentage } from './examples/number-with-percentage'
 export { NumberWithUnit } from './examples/number-with-unit'
+
+export { RelativeTimeBasic } from './examples/relative-time-basic'
+export { RelativeTimeShort } from './examples/relative-time-short'