+++ packages/solid/src/components/format/index.tsx
@@ -1,4 +1,5 @@
 export { FormatByte, type FormatByteProps } from './format-byte'
 export { FormatNumber, type FormatNumberProps } from './format-number'
+export { FormatRelativeTime, type FormatRelativeTimeProps } from './format-relative-time'
 
 export * as Format from './format'