+++ packages/vue/src/components/format/format.stories.vue
@@ -10,6 +10,8 @@ import NumberWithCurrency from './examples/number-with-currency.vue'
 import NumberWithLocale from './examples/number-with-locale.vue'
 import NumberWithPercentage from './examples/number-with-percentage.vue'
 import NumberWithUnit from './examples/number-with-unit.vue'
+import RelativeTimeBasic from './examples/relative-time-basic.vue'
+import RelativeTimeShort from './examples/relative-time-short.vue'
 </script>
 
 <template>
@@ -29,6 +31,7 @@ import NumberWithUnit from './examples/number-with-unit.vue'
     <Variant title="Number with Compact">
       <NumberWithCompact />
     </Variant>
+
     <Variant title="Byte Basic">
       <ByteBasic />
     </Variant>
@@ -46,6 +49,13 @@ import NumberWithUnit from './examples/number-with-unit.vue'
     </Variant>
     <Variant title="Number with Currency">
       <NumberWithCurrency />
+    </Variant>
+
+    <Variant title="Relative Time Basic">
+      <RelativeTimeBasic />
+    </Variant>
+    <Variant title="Relative Time Short">
+      <RelativeTimeShort />
     </Variant>
   </Story>
 </template>